requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
dj-database-url==2.2.0
Django==5.0.11
django-cors-headers==4.3.1
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.1
sqlparse==0.5.0
typing_extensions==4.12.1
uvicorn[standard]==0.30.1
